Bugfix
Line 294 RegExp needs to have a '+' added after the [0-9] as the quantity += will not properly add to total quantity.
Ex: Have a user grab 11 quantity of an item and then grab another 12 right after and you will get 113 quantity instead of the proper 23. Other issues will result in NaN inserted in the quantity because of this.